+++ enjoy.py
@@ -1,6 +1,7 @@
 import argparse
 import importlib
 import os
+import sys
 
 import numpy as np
 import torch as th
@@ -140,7 +141,19 @@ def main():  # noqa: C901
         # Dummy buffer size as we don't need memory to enjoy the trained agent
         kwargs.update(dict(buffer_size=1))
 
-    model = ALGOS[algo].load(model_path, env=env, **kwargs)
+    # Check if we are running python 3.8+
+    # we need to patch saved model under python 3.6/3.7 to load them
+    newer_python_version = sys.version_info.major == 3 and sys.version_info.minor >= 8
+
+    custom_objects = {}
+    if newer_python_version:
+        custom_objects = {
+            "learning_rate": 0.0,
+            "lr_schedule": lambda _: 0.0,
+            "clip_range": lambda _: 0.0,
+        }
+
+    model = ALGOS[algo].load(model_path, env=env, custom_objects=custom_objects, **kwargs)
 
     obs = env.reset()
 
@@ -197,7 +210,7 @@ def main():  # noqa: C901
 
     except KeyboardInterrupt:
         pass
-        
+
     if args.verbose > 0 and len(successes) > 0:
         print(f"Success rate: {100 * np.mean(successes):.2f}%")